+++ src/app/models/__init__.py
@@ -9,7 +9,7 @@ from .choice_option import ChoiceOption
 from .form_field import FormField
 from .setting_autocomplete import SettingAutocomplete
 from .shared_field import SharedField
-from .range import Range
+from .range import Range, RangeSchema
 from .form_result import FormResult
 from .field_range import FieldRange
 from .group_user import GroupUser
